move readDirectoryTree into process_colourtables.cpp, as that's where it's used
