+++ buildtools/CMakeLists.txt
@@ -6,7 +6,7 @@ endif()
 # Permit #include <morph/header.h>
 include_directories(BEFORE ${PROJECT_SOURCE_DIR})
 if(NOT WIN32)
-  # Colourmap processing (uses readDirectoryTree which is not Windows compatible)
+  # Colourmap processing is not Windows compatible
   add_executable(process_colourmap process_colourmap.cpp)
   add_executable(process_colourtables process_colourtables.cpp)
 endif()
